# docs/openapi-engine-live.yaml
openapi: 3.0.1
info:
  title: APIGatewayMQ Engine API
  description: 交易引擎：下單、風控、持倉、PnL、成交查詢（經 Gateway 代理或直連）
  version: v1
servers:
- url: http://localhost
  description: Generated server url
tags:
- name: Jobs
  description: 排程 Job（JOB-A 超時取消 / JOB-B PnL 快照 / JOB-C 失敗重試 / JOB-D 資料清理）
paths:
  /api/v1/orders:
    get:
      tags:
      - order-controller
      operationId: listOrders_1
      parameters:
      - name: symbol
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - NEW
          - PARTIALLY_FILLED
          - FILLED
          - REJECTED
          - CANCELLED
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: clientOrderId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/OrderResponse'
                - $ref: '#/components/schemas/PagedResponseOrderResponse'
    post:
      tags:
      - order-controller
      operationId: placeOrder
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /api/v1/orders/{orderId}/fill:
    post:
      tags:
      - order-controller
      operationId: completeFill
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /api/v1/jobs/stale-order-cancellation:
    post:
      tags:
      - Jobs
      summary: JOB-A 手動執行：取消逾時未成交訂單
      operationId: runStaleOrderCancellation
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobRunResponse'
  /api/v1/jobs/pnl-snapshot:
    post:
      tags:
      - Jobs
      summary: JOB-B 手動執行：建立當日 PnL/持倉結算快照
      operationId: runPnlSnapshot
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobRunResponse'
  /api/v1/jobs/failed-command-retry:
    post:
      tags:
      - Jobs
      summary: JOB-C 手動執行：重試失敗的下單指令
      operationId: runFailedCommandRetry
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobRunResponse'
  /api/v1/jobs/cleanup:
    post:
      tags:
      - Jobs
      summary: JOB-D 手動執行：清理過期審計事件與失敗指令
      operationId: runCleanup
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobRunResponse'
  /api/v1/orders/{orderId}/cancel:
    patch:
      tags:
      - order-controller
      operationId: cancelOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /api/v1/trades:
    get:
      tags:
      - trade-controller
      operationId: listTrades
      parameters:
      - name: orderId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedResponseTradeDetailResponse'
  /api/v1/trades/{tradeId}:
    get:
      tags:
      - trade-controller
      operationId: getTrade
      parameters:
      - name: tradeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TradeDetailResponse'
  /api/v1/positions:
    get:
      tags:
      - position-controller
      operationId: listPositions
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedResponsePositionResponse'
  /api/v1/positions/{symbol}:
    get:
      tags:
      - position-controller
      operationId: getPosition
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PositionResponse'
  /api/v1/pnl:
    get:
      tags:
      - pn-l-controller
      operationId: getPnL
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PnLResponse'
  /api/v1/pnl-snapshots:
    get:
      tags:
      - Jobs
      summary: 查詢 PnL 結算快照（JOB-B 輸出），未指定日期則為今日
      operationId: listPnlSnapshots
      parameters:
      - name: date
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PnlSnapshotResponse'
  /api/v1/orders/{orderId}:
    get:
      tags:
      - order-controller
      operationId: getOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /api/v1/orders/{orderId}/events:
    get:
      tags:
      - order-controller
      operationId: getOrderEvents
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderEventsResponse'
  /api/v1/failed-commands:
    get:
      tags:
      - Jobs
      summary: 查詢失敗指令 DLQ（JOB-C 佇列），可依狀態過濾
      operationId: listFailedCommands
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - PENDING
          - SUCCEEDED
          - DEAD
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FailedCommandResponse'
components:
  schemas:
    CreateOrderRequest:
      required:
      - price
      - quantity
      - side
      - symbol
      type: object
      properties:
        clientOrderId:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum:
          - BUY
          - SELL
        quantity:
          type: number
        price:
          type: number
    OrderResponse:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
        clientOrderId:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum:
          - BUY
          - SELL
        quantity:
          type: number
        filledQuantity:
          type: number
        price:
          type: number
        status:
          type: string
          enum:
          - NEW
          - PARTIALLY_FILLED
          - FILLED
          - REJECTED
          - CANCELLED
        rejectReason:
          type: string
        riskRuleCode:
          type: string
        trade:
          $ref: '#/components/schemas/TradeResponse'
        trades:
          type: array
          items:
            $ref: '#/components/schemas/TradeResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TradeResponse:
      type: object
      properties:
        executedPrice:
          type: number
        executedQty:
          type: number
        executedAt:
          type: string
          format: date-time
    JobRunResponse:
      type: object
      properties:
        job:
          type: string
        affected:
          type: integer
          format: int64
        detail:
          type: string
        executedAt:
          type: string
          format: date-time
    PageMeta:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PagedResponseTradeDetailResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TradeDetailResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'
    TradeDetailResponse:
      type: object
      properties:
        tradeId:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        executedPrice:
          type: number
        executedQty:
          type: number
        executedAt:
          type: string
          format: date-time
    PagedResponsePositionResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PositionResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'
    PositionResponse:
      type: object
      properties:
        symbol:
          type: string
        quantity:
          type: number
        avgPrice:
          type: number
        unrealizedPnl:
          type: number
        updatedAt:
          type: string
          format: date-time
    PnLResponse:
      type: object
      properties:
        totalUnrealizedPnl:
          type: number
        positions:
          type: array
          items:
            $ref: '#/components/schemas/PositionPnlItem'
        asOf:
          type: string
          format: date-time
    PositionPnlItem:
      type: object
      properties:
        symbol:
          type: string
        unrealizedPnl:
          type: number
    PnlSnapshotResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        snapshotDate:
          type: string
          format: date
        symbol:
          type: string
        quantity:
          type: number
        avgPrice:
          type: number
        markPrice:
          type: number
        unrealizedPnl:
          type: number
        createdAt:
          type: string
          format: date-time
    PagedResponseOrderResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'
    EventItem:
      type: object
      properties:
        event:
          type: string
          enum:
          - RECEIVED
          - RISK_CHECK
          - APPROVED
          - REJECTED
          - PARTIALLY_FILLED
          - FILLED
          - CANCELLED
          - OUTCOME_RECORDED
          - DISCIPLINE_FLAG
          - POSITION_UPDATED
        riskRuleCode:
          type: string
        rejectReason:
          type: string
        payloadJson:
          type: string
        createdAt:
          type: string
          format: date-time
    OrderEventsResponse:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventItem'
    FailedCommandResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        commandId:
          type: string
        clientOrderId:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum:
          - BUY
          - SELL
        quantity:
          type: number
        price:
          type: number
        attempts:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - PENDING
          - SUCCEEDED
          - DEAD
        failureReason:
          type: string
        nextRetryAt:
          type: string
          format: date-time